HW-5 : Move ui classes
