+++ class15/showmiso/ArchitectureStudy/app/src/main/res/layout/activity_main.xml
@@ -11,13 +11,13 @@
 
         <variable
             name="activity"
-            type="com.showmiso.architecturestudy.MainActivity" />
+            type="com.showmiso.architecturestudy.ui.MainActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
 
         <androidx.appcompat.widget.LinearLayoutCompat
             android:id="@+id/layout_input"
